requirements.txt
fastapi==0.112.0
psycopg==3.2.1
pydantic==2.8.2
PyYAML==6.0.1
Requests==2.32.3
